Rename based on new naming convention
