+++ services/npm/npm-license.service.js
